\name{winsorize-package}
\alias{winsorize-package}
\docType{package}
\title{
  Winsorize Data
}
\description{
  Remove outliers by means of winsorization, ie shrinking outlying
  observations to the border of the main part of the data.
}
\details{
  This package started from the excellent robustHD package by Andreas
  Alfons in order to reduce the number dependent package being pulled
  in. We expect to update the code over time.
}
\author{
  Andreas Alfons wrote \code{winsorize} as part as his excellent
  \code{\link[robustHD:robustHD-package]{robustHD}} package.

  Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\keyword{ package }
\seealso{
  See the \code{\link[robustHD:robustHD-package]{robustHD}} package for more.
}
